vauth: Moved the NTLM authentication code to the new vauth directory
