+++ lib/Makefile.vc6
@@ -542,11 +542,9 @@ X_OBJS= \
 	$(DIROBJ)\curl_multibyte.obj \
 	$(DIROBJ)\curl_ntlm.obj \
 	$(DIROBJ)\curl_ntlm_core.obj \
-	$(DIROBJ)\curl_ntlm_msgs.obj \
 	$(DIROBJ)\curl_ntlm_wb.obj \
 	$(DIROBJ)\curl_rtmp.obj \
 	$(DIROBJ)\curl_sasl.obj \
-	$(DIROBJ)\curl_sasl_sspi.obj \
 	$(DIROBJ)\curl_sspi.obj \
 	$(DIROBJ)\curl_threads.obj \
 	$(DIROBJ)\cyassl.obj \
@@ -626,6 +624,8 @@ X_OBJS= \
 	$(DIROBJ)\digest_sspi.obj \
 	$(DIROBJ)\krb5_gssapi.obj \
 	$(DIROBJ)\krb5_sspi.obj \
+	$(DIROBJ)\ntlm.obj \
+	$(DIROBJ)\ntlm_sspi.obj \
 	$(DIROBJ)\vtls.obj \
 	$(DIROBJ)\openssl.obj \
 	$(DIROBJ)\strdup.obj \
